Cleaning:move the watcher and clean blockchain
