+++ core/watcher_test.go
@@ -1,4 +1,4 @@
-package blockchain
+package core
 
 import (
 	"testing"
